[Refactoring] - manda Status, TopBar 분리
